Fix escaping and remove some incorrect documentation
It turns out some of the docs from the previous commit were just wrong,
and most things can be done with the closure macro.  Remove the wrong
bits, and keep some useful refactoring.  Also move around some of the
escaping, as it seems it can be done in a simpler way.